Move new quiz methods back to convert file
This isn't reused or tested separately, so just put it in the main convert file so it's easier to navigate around in
